# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenPlugin",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/function-providers": {
            "get": {
                "tags": [
                    "function-providers"
                ],
                "summary": "Get Function Providers",
                "description": "Enpoint to retrieve list of available function providers",
                "operationId": "get_function_providers_api_function_providers_get",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "title": "Type"
                        }
                    },
                    {
                        "name": "openai_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Openai Api Key"
                        }
                    },
                    {
                        "name": "anthropic_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Anthropic Api Key"
                        }
                    },
                    {
                        "name": "gemini_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Gemini Api Key"
                        }
                    },
                    {
                        "name": "cohere_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Cohere Api Key"
                        }
                    },
                    {
                        "name": "fireworks_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Fireworks Api Key"
                        }
                    },
                    {
                        "name": "mistral_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Mistral Api Key"
                        }
                    },
                    {
                        "name": "together_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Together Api Key"
                        }
                    },
                    {
                        "name": "groq_api_key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Groq Api Key"
                        }
                    },
                    {
                        "name": "openplugin_manifest_url",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Openplugin Manifest Url"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FunctionProvider"
                                    },
                                    "title": "Response Get Function Providers Api Function Providers Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/function-provider-request": {
            "get": {
                "tags": [
                    "function-provider-request"
                ],
                "summary": "Get Function Provider Request",
                "description": "Enpoint to retrieve function provider request JSON",
                "operationId": "get_function_provider_request_api_function_provider_request_get",
                "parameters": [
                    {
                        "name": "function_provider_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Function Provider Name"
                        }
                    },
                    {
                        "name": "openplugin_manifest_url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Openplugin Manifest Url"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FunctionProviderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/run-function-provider": {
            "post": {
                "tags": [
                    "run-function-provider"
                ],
                "summary": "Run Function Provider",
                "description": "Enpoint to run a function provider",
                "operationId": "run_function_provider_api_run_function_provider_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RunFunctionInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FunctionProviderRunResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/plugin-validator": {
            "post": {
                "tags": [
                    "plugin-validator"
                ],
                "summary": "Plugin Execution Pipeline",
                "description": "Enpoint to validate a plugin manifest",
                "operationId": "plugin_execution_pipeline_api_plugin_validator_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "title": "Openplugin Manifest Obj"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PluginValidationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/api/openapi-param-parser": {
            "get": {
                "tags": [
                    "openapi-param-parse"
                ],
                "summary": "Openapi Param Parser",
                "description": "Enpoint to parse OpenAPI doc and extract functions",
                "operationId": "openapi_param_parser_api_openapi_param_parser_get",
                "parameters": [
                    {
                        "name": "openapi_doc_url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Openapi Doc Url"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FunctionResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Config": {
                "properties": {
                    "provider": {
                        "type": "string",
                        "title": "Provider",
                        "default": "openai"
                    },
                    "openai_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Openai Api Key"
                    },
                    "cohere_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cohere Api Key"
                    },
                    "mistral_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mistral Api Key"
                    },
                    "fireworks_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fireworks Api Key"
                    },
                    "google_palm_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Google Palm Key"
                    },
                    "gemini_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gemini Api Key"
                    },
                    "anthropic_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Anthropic Api Key"
                    },
                    "together_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Together Api Key"
                    },
                    "aws_access_key_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aws Access Key Id"
                    },
                    "aws_secret_access_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aws Secret Access Key"
                    },
                    "aws_region_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aws Region Name"
                    },
                    "azure_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Azure Api Key"
                    },
                    "groq_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Groq Api Key"
                    }
                },
                "type": "object",
                "title": "Config",
                "description": "Represents the API configuration for a plugin."
            },
            "FunctionProvider": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "required_auth_keys": {
                        "items": {},
                        "type": "array",
                        "uniqueItems": true,
                        "title": "Required Auth Keys"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "is_supported": {
                        "type": "boolean",
                        "title": "Is Supported",
                        "default": false
                    },
                    "is_default": {
                        "type": "boolean",
                        "title": "Is Default",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "required_auth_keys",
                    "type"
                ],
                "title": "FunctionProvider"
            },
...
            "FunctionProviderResponse": {
                "properties": {
                    "fc_request_json": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Fc Request Json"
                    }
                },
                "type": "object",
                "required": [
                    "fc_request_json"
                ],
                "title": "FunctionProviderResponse"
            },
            "FunctionProviderRunResponse": {
                "properties": {
                    "fc_response_json": {
                        "type": "object",
                        "title": "Fc Response Json"
                    }
                },
                "type": "object",
                "required": [
                    "fc_response_json"
                ],
                "title": "FunctionProviderRunResponse"
            },
            "FunctionResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "functions": {
                        "type": "object",
                        "title": "Functions"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "functions"
                ],
                "title": "FunctionResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PluginValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "plugin_name": {
                        "type": "string",
                        "title": "Plugin Name"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "plugin_name"
                ],
                "title": "PluginValidationResponse"
            },
...
            "RunFunctionInput": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "title": "Prompt"
                    },
                    "function_provider_name": {
                        "type": "string",
                        "title": "Function Provider Name"
                    },
                    "openplugin_manifest_url": {
                        "type": "string",
                        "title": "Openplugin Manifest Url"
                    },
                    "config": {
                        "$ref": "#/components/schemas/Config"
                    },
                    "function_json": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Function Json"
                    }
                },
                "type": "object",
                "required": [
                    "prompt",
                    "function_provider_name",
                    "openplugin_manifest_url",
                    "config"
                ],
                "title": "RunFunctionInput"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        }
    }
}